finble/requirements.txt
asgiref==3.5.2
Django==3.0.8
django-cors-headers==3.4.0
django-environ==0.9.0
PyMySQL==1.0.2
PyJWT==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
sqlparse==0.4.3
gunicorn==20.1.0
